+++ fixtures/mtls/gen_test_certs.sh
@@ -1,4 +1,7 @@
 #!/usr/bin/env bash
+# Copyright (c) HashiCorp, Inc.
+# SPDX-License-Identifier: MPL-2.0
+
 
 GOOS=`go env GOOS`
 GOARCH=`go env GOARCH`