+++ internal/cli/run.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package cli
 
 import (
@@ -66,7 +69,14 @@ func (c *RunCommand) run() int {
 
 	// Render the pack now, before creating the deployer. If we get an error
 	// we won't make it to the deployer.
-	r, err := renderPack(packManager, c.baseCommand.ui, errorContext)
+	r, err := renderPack(
+		packManager,
+		c.baseCommand.ui,
+		false,
+		false,
+		c.baseCommand.ignoreMissingVars,
+		errorContext,
+	)
 	if err != nil {
 		return 255
 	}