+++ internal/pkg/cache/pack.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package cache
 
 import (
@@ -70,10 +73,6 @@ func invalidPackDefinition(provider cacheOperationProvider) *Pack {
 		Ref: provider.AtRef(),
 		Pack: &pack.Pack{
 			Metadata: &pack.Metadata{
-				App: &pack.MetadataApp{
-					URL:    "",
-					Author: "",
-				},
 				Pack: &pack.MetadataPack{
 					Name:        provider.ForPackName(),
 					Description: "",