+++ sdk/pack/metadata_test.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package pack
 
 import (
@@ -15,8 +18,7 @@ func TestMetadata_ConvertToMapInterface(t *testing.T) {
 		{
 			inputMetadata: &Metadata{
 				App: &MetadataApp{
-					URL:    "https://example.com",
-					Author: "Timothy J. Berners-Lee",
+					URL: "https://example.com",
 				},
 				Pack: &MetadataPack{
 					Name:        "Example",
@@ -28,8 +30,7 @@ func TestMetadata_ConvertToMapInterface(t *testing.T) {
 			expectedOutput: map[string]interface{}{
 				"nomad_pack": map[string]interface{}{
 					"app": map[string]interface{}{
-						"url":    "https://example.com",
-						"author": "Timothy J. Berners-Lee",
+						"url": "https://example.com",
 					},
 					"pack": map[string]interface{}{
 						"name":        "Example",
@@ -55,8 +56,7 @@ func TestMetadata_ConvertToMapInterface(t *testing.T) {
 			expectedOutput: map[string]interface{}{
 				"nomad_pack": map[string]interface{}{
 					"app": map[string]interface{}{
-						"url":    "https://example.com",
-						"author": "",
+						"url": "https://example.com",
 					},
 					"pack": map[string]interface{}{
 						"name":        "Example",
@@ -67,6 +67,26 @@ func TestMetadata_ConvertToMapInterface(t *testing.T) {
 				},
 			},
 			name: "some metadata values populated",
+		},
+		{
+			inputMetadata: &Metadata{
+				App: &MetadataApp{
+					URL:    "https://example.com",
+					Author: "The Nomad Team",
+				},
+				Pack: &MetadataPack{},
+			},
+			expectedOutput: map[string]interface{}{
+				"nomad_pack": map[string]interface{}{
+					"app": map[string]interface{}{
+						"url": "https://example.com",
+					},
+					"pack": map[string]interface{}{"name": "", "description": "", "url": "", "version": ""},
+				},
+			},
+			// TODO test added to cover graceful failure while we're in the process of
+			// retiring "Author" metadata field. Can be removed later.
+			name: "author field ignored gracefully",
 		},
 	}
 
@@ -85,8 +105,7 @@ func TestMetadata_Validate(t *testing.T) {
 		{
 			inputMetadata: &Metadata{
 				App: &MetadataApp{
-					URL:    "https://example.com",
-					Author: "Timothy J. Berners-Lee",
+					URL: "https://example.com",
 				},
 				Pack: &MetadataPack{
 					Name:        "Example",