+++ terminal/doc.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 // Package terminal is used by plugins to read and write to a terminal
 // UI. The abstractions presented in this package are meant to be portable
 // across a variety of styles that may be presented to the user and enables